requirements.txt
flask==2.3.3
Pillow==10.0.0
python-dotenv==1.0.0
